LayoutTests/fast/dom/Geolocation/disconnected-frame-permission-denied-expected.txt

frame "<!--frame1-->" - has 1 onunload handler(s)
Tests that when a request is made on a Geolocation object, permission is denied and its Frame is disconnected before a callback is made, the error callback is invoked with PERMISSION_DENIED.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS successCallbackInvoked is false
PASS firstPositionError.code is GeolocationPositionError.PERMISSION_DENIED
PASS firstPositionError.message is 'User denied Geolocation'
PASS secondPositionError.code is GeolocationPositionError.PERMISSION_DENIED
PASS secondPositionError.message is 'User denied Geolocation'
PASS successfullyParsed is true

TEST COMPLETE
